Organized geo ref stuff for newness
