+++ Runtime/Prefabs/ROS/ROSClock.prefab
@@ -10,9 +10,8 @@ GameObject:
   m_Component:
   - component: {fileID: 2970021653359921310}
   - component: {fileID: 1662493632097492073}
-  - component: {fileID: 3062139818935489222}
   m_Layer: 0
-  m_Name: ROSSingletons
+  m_Name: ROSClock
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -48,16 +47,3 @@ MonoBehaviour:
   m_ClockMode: 0
   m_LastSetClockMode: 0
   m_PublishRateHz: 100
---- !u!114 &3062139818935489222
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6357952754565123669}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ee81888241533583a9d98e158a029d69, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  frequency: 1
