+++ Runtime/Scripts/GeoRef.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4eeb34ec464ed274f951463899ca96a6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
